+++ app/src/main/java/com/numero/material_gallery/components/list/ComponentListActivity.kt
@@ -37,6 +37,7 @@ import com.numero.material_gallery.components.sheet.ModalBottomSheetActivity
 import com.numero.material_gallery.components.slider.SliderActivity
 import com.numero.material_gallery.components.snackbar.SnackbarActivity
 import com.numero.material_gallery.components.tab.TabActivity
+import com.numero.material_gallery.components.textfield.TextFieldActivity
 import com.numero.material_gallery.view.ListItemAdapter
 import kotlinx.android.synthetic.main.activity_main.*
 import kotlinx.android.synthetic.main.content_main.*
