+++ app/src/main/java/com/numero/material_gallery/components/textfield/TextFieldActivity.kt
@@ -1,4 +1,4 @@
-package com.numero.material_gallery.activity
+package com.numero.material_gallery.components.textfield
 
 import android.content.Context
 import android.content.Intent
